refactor: add bugetInput file
